// server/src/logical/application/dto/create-application.dto.ts
export class CreateApplicationDto {
  name: string;
  sex: string;
  birthday: Date;
  passport: string;
  phone: number;
  record: string;
  // university: string;
  // profess: string;
  // work: string;
  project: string;
  serve_duration: number;
  languages: string;
  skills: string;
  hobbies: string;
}